requirements.txt
# process pdf
pandas
numpy
tabula-py

# API
requests

# DEV
jupyterlab
